Data Restructure and Updates for Filter Functionlity, state moved to data file filterOption and guestsLists
