+++ src/data/guestsLists.js
@@ -1,0 +1,54 @@
+import { v4 as uuidv4 } from 'uuid';
+
+const guestsList = [
+    {
+        guestName: "S Kraswo",
+        noOfGuests: 2,
+        checkIn: "13 Jul 2022",
+        checkout: "16 Jul 2022",
+        room: "Single Room",
+        bookedOn: "10 Jul 2022",
+        status: "Ok",
+        price: 135,
+        commission: 16.20,
+        bookingNumber: uuidv4()
+    },
+    {
+        guestName: "S Kraswo",
+        noOfGuests: 2,
+        checkIn: "13 Jul 2022",
+        checkout: "16 Jul 2022",
+        room: "Single Room",
+        bookedOn: "10 Jul 2022",
+        status: "Ok",
+        price: 135,
+        commission: 16.20,
+        bookingNumber: uuidv4()
+    },
+    {
+        guestName: "S Kraswo",
+        noOfGuests: 2,
+        checkIn: "13 Jul 2022",
+        checkout: "16 Jul 2022",
+        room: "Single Room",
+        bookedOn: "10 Jul 2022",
+        status: "Ok",
+        price: 135,
+        commission: 16.20,
+        bookingNumber: uuidv4()
+    },
+    {
+        guestName: "S Kraswo",
+        noOfGuests: 2,
+        checkIn: "13 Jul 2022",
+        checkout: "16 Jul 2022",
+        room: "Single Room",
+        bookedOn: "10 Jul 2022",
+        status: "Ok",
+        price: 135,
+        commission: 16.20,
+        bookingNumber: uuidv4()
+    },
+]
+
+export default guestsList;
