+++ src/pages/Reservations/index.js
@@ -1,6 +1,7 @@
 import React, {useState} from 'react';
-import { v4 as uuidv4 } from 'uuid';
 import DatePicker from "react-datepicker";
+import filterOptions from '../../data/filterOptions';
+import guestsList from '../../data/guestsLists';
 
 import "react-datepicker/dist/react-datepicker.css";
 import './index.scss';
@@ -14,145 +15,22 @@ import InputSearch from '../../components/InputSearch';
 import Table from '../../components/Table';
 
 const Reservations = () => {
+
     const [filter, setFilter] = useState({
-        reservationType: [
-            {
-                id: uuidv4(),
-                value: "reservation",
-                name: "Reservation"
-            },
-            {
-                id: uuidv4(),
-                value: "checkIn",
-                name: "Check-In"
-            },
-            {
-                id: uuidv4(),
-                value: "checkOut",
-                name: "Check-Out"
-            },
-            {
-                id: uuidv4(),
-                value: "invoice",
-                name: "Invoice"
-            },
-            {
-                id: uuidv4(),
-                value: "checkOut",
-                name: "Stay"
-            },
-        ],
+        dateOf: '--',
         dateFrom: new Date(),
         dateTo: new Date(),
-        moreFilter: {
-            reservationStatus: [
-                    {
-                        id: uuidv4(),
-                        value: "ok",
-                        name: "Ok",
-                    },
-                    {
-                        id: uuidv4(),
-                        value: "cancelled",
-                        name: "Cancelled",
-                    }, 
-                    {
-                        id: uuidv4(),
-                        value: "noShow",
-                        name: "No show",
-                    },
-                    {
-                        id: uuidv4(),
-                        value: "corporateCard",
-                        name: "Corporate Card",
-                    }
-            ],
-            guestCommunication:  [
-                    {
-                        id: uuidv4(),
-                        value: "pendingRequest",
-                        name: "Pending guest request",
-                    },
-                    {
-                        id: uuidv4(),
-                        value: "invoiceRequired",
-                        name: "Invoice required",
-                    }
-            ],
-            invalidCreditCard: [
-                    {
-                        id: uuidv4(),
-                        value: "updated",
-                        name: "Updated",
-                    },
-                    {
-                        id: uuidv4(),
-                        value: "pending",
-                        name: "Pending",
-                    }
-            ] 
-        }
     })
-
-    const [guestsList] = useState([
-        {
-            guestName: "S Kraswo",
-            noOfGuests: 2,
-            checkIn: "13 Jul 2022",
-            checkout: "16 Jul 2022",
-            room: "Single Room",
-            bookedOn: "10 Jul 2022",
-            status: "Ok",
-            price: 135,
-            commission: 16.20,
-            bookingNumber: uuidv4()
-        },
-        {
-            guestName: "S Kraswo",
-            noOfGuests: 2,
-            checkIn: "13 Jul 2022",
-            checkout: "16 Jul 2022",
-            room: "Single Room",
-            bookedOn: "10 Jul 2022",
-            status: "Ok",
-            price: 135,
-            commission: 16.20,
-            bookingNumber: uuidv4()
-        },
-        {
-            guestName: "S Kraswo",
-            noOfGuests: 2,
-            checkIn: "13 Jul 2022",
-            checkout: "16 Jul 2022",
-            room: "Single Room",
-            bookedOn: "10 Jul 2022",
-            status: "Ok",
-            price: 135,
-            commission: 16.20,
-            bookingNumber: uuidv4()
-        },
-        {
-            guestName: "S Kraswo",
-            noOfGuests: 2,
-            checkIn: "13 Jul 2022",
-            checkout: "16 Jul 2022",
-            room: "Single Room",
-            bookedOn: "10 Jul 2022",
-            status: "Ok",
-            price: 135,
-            commission: 16.20,
-            bookingNumber: uuidv4()
-        },
-    ])
 
     const [moreFilterStatus, setMoreFilterStatus] = useState(false);
 
     const moreFilterDropdown = () => {
         setMoreFilterStatus(!moreFilterStatus)
-        console.log("Dropdown Clicked")
     }
 
-    console.log(filter.dateFrom)
+    const handleChange = (e) => {
+        setFilter({...filter, dateOf: e.target.value })
+    }
 
   return (
     <div className='Reservations CustomContainer'>
@@ -162,7 +40,7 @@ const Reservations = () => {
         {/* Filter Section */}
         <section className='Filter'>
             <div className="FormGroup">
-                <InputSelect id="dateOf" label="Date of" options={filter.reservationType} />
+                <InputSelect id="dateOf" label="Date of" options={filterOptions.reservationType} value={filter.dateOf} handleChange={handleChange} />
             </div>
             <div className="FormGroup">
                 <label htmlFor="dateFrom">From</label>
@@ -180,17 +58,18 @@ const Reservations = () => {
                 <button className='Btn BtnBlue'>Show</button>
             </div>
         </section>
-
+        
+        {/* More Filters Section */}
         <section className="MoreFilters" style={{maxHeight: moreFilterStatus ? '500px' : '0'}}>
             <div className='FilterWrapper'>
                 <div className="FormGroup">
-                    <InputCheckBox name="reservationStatus" title="Reservation Status" options={filter.moreFilter.reservationStatus} />
+                    <InputCheckBox name="reservationStatus" title="Reservation Status" options={filterOptions.moreFilter.reservationStatus} />
                 </div>
                 <div className="FormGroup">
-                    <InputCheckBox name="guestCommunication" title="Guest Communication" options={filter.moreFilter.guestCommunication} />
+                    <InputCheckBox name="guestCommunication" title="Guest Communication" options={filterOptions.moreFilter.guestCommunication} />
                 </div>
                 <div className="FormGroup">
-                    <InputCheckBox name="invalidCreditCard" title="Invalid Credit Card" options={filter.moreFilter.invalidCreditCard} />
+                    <InputCheckBox name="invalidCreditCard" title="Invalid Credit Card" options={filterOptions.moreFilter.invalidCreditCard} />
                 </div>
 
                 <div className="FormGroup">
@@ -200,12 +79,10 @@ const Reservations = () => {
             </div>
         </section>
         
-
         {/* Reservation User List Section */}
         <section className="UserList">
             <Table guestsList={guestsList} />
         </section>
-        
 
     </div>
   );
